Refactor dataset selection components and remove BenchmarkSelector
- Removed BenchmarkSelector.vue and its associated logic from CustomLayerSwitcher.vue and AnalysisPanel.vue.
- Introduced CompactDatasetSelector.vue to streamline dataset selection with a more compact UI.
- Updated AnalysisPanel.vue to utilize CompactDatasetSelector for dataset selection.
- Adjusted localization files to include new strings for dataset selection.
- Cleaned up styles and improved user experience for dataset selection dialogs.
